Add files via upload
Files for having functioning study group database on the courses page of the website. 
Also- I updated the index.html file, which is the login/landing page. URGroup ---> StudyBuddy. 

Need: to check that the PHP script works on the courses.html page to display data from the groupsDB. 
Need: update and delete features.